+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This document outlines major changes between releases.
 - Support for legacy access boxes carrying session token v1 and bearer tokens (#1271)
 - Backward-compatibility code for legacy bucket settings, CORS, tags and notification meta objects (#1265)
 - Object ACL support (#1275)
+- Authmate `bearer-rules` option (#1306)
 
 ### Upgrading from 0.44.0